+++ cognite/experimental/_api/relationships.py
@@ -3,6 +3,7 @@ from typing import Any, Dict, Generator, List, Optional, Set, Union
 
 from cognite.client import utils
 from cognite.client._api_client import APIClient
+
 from cognite.experimental.data_classes import Relationship, RelationshipFilter, RelationshipList