+++ CHANGELOG.md
@@ -12,6 +12,24 @@ Changes are grouped as follows
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities.
 
+## [0.33.0] - 2020-12-02
+### Added
+- Method `get_input_data` on the `FunctionSchedule`-class. 
+- Method `get_input_data` on the `FunctionScheduleAPI`-class.
+
+## [0.32.0] - 2020-12-03
+### Fixed
+- Fix the client initialization to correctly pass all keyword arguments.
+
+## [0.31.2] - 2020-12-03
+### Fixed
+- Fix the filter functionality of `annotations`
+
+## [0.31.1] - 2020-11-30
+### Added
+- Argument `function_call_info` to the definition of the `handle` function. This argument is a dictionary with keys `function_id` and, if the call is scheduled, `schedule_id` and `scheduled_time`.
+- Property `scheduled_time` on `FunctionCall` class.
+
 ## [0.31.0] - 2020-11-25
 ### Changed
 - Core entity matching moved to beta client, experimental now extends this.